Remove old validators and some more extractors
Note, there is some work to fix GA stuff and also improve nav
handling which we will need to look at later on.
